Fix bug with event api deletion
Issue: When you delete an event using the event api, it was incorrectly
returning "Event ID ${eventid} not found" when the event was found and deleted.
It was returning the same message if the event was not found.

Delete now returns properly on file write error, event not found and successful
deletion.