Reports only in cluster view and not host view
When collection the optional graphs for a host, the logic is as follows:

- get default_reports from default.json
- get cluster_override_reports from cluster_<name>.json
- if host_<name>.json exists, take these as overwrite_reports, otherwise
  use the cluster_override
- at last merge all three reports

So, even if there is a host_<name>.json, we always merge in the
cluster reports.

For the Power systems we use, we have graphs that are only useful in
the cluster view. These cluster graphs should not be displayed for the
hosts. Since the current behavior might be useful for other people
make it configurable with $conf['optional_cluster_graphs_for_host_view']