We have lots of json-reports for clusters that we generate with a script. At the same time we recreate the cluster_<name>.json file and add these reports to the included_reports. This works quite well, but means that there can't be local changes in the includes/excludes reports.
So we've patched cluster_view.php, to look for reports matching
cluster_<name>_.*_report.json (or php). If such a report exists, it is
automatically added to the included_reports array and gets displayed.

Since the change is very easy to port to the host_view (and we also
have lot's of reports for hosts as well), that hunk is included too.

Since this is a very visible change, review is welcome. If the changes
are merged a description should be added to some documentation and
possibly the ganglia book as well.