hide automatic graphs from view in edit_optional_graphs and metric drop-down list
In the cluster view we can select the metric that is displayed
for each host.  Until now all reports/metrics are displayed in
the dropdown-list "Metric".

In the "Edit Optional Graphs" we had all reports for cluster and host.

In our system we have dozens of reports for our clusters (but only
some relevant for each cluster) and hundreds of different reports for
hosts (and again only some relevant for each host).

The earlier patches added the reports automagically to clusters and
hosts respectively, so there is no need to add these reports in the
edit_optional_graphs.php and the metric drop-down list in the
cluster_view.

For both lists we exclude the (cluster|host)_*_report.json from
view, because they are always included as we need it.