hide selected metrics from drop down menu
We have hundreds of local metrics that are only useful for a single
host. Currently all of these metrics are added to the drop down
menu in the cluster view.

This patch defines the configuration option
'cluster_hide_metrics_from_menu', a regular expression. All metrics
matching this regular expression are hidden from the drop down menu.