Added mem_free to graph.d/mem_report.php graph. Was present in ganglia web 1. Also ported RGB color definition from there.
Reason:
Displaying memory graph for linux works as expected, but metrics from windows
Host sFlow agents displayed swap usage where free memory should be displayed.
Therefore it presented itself as when machine has all of it's memory and lots
of swap space occupied, when that was clearly not the case.

This change adds it back to graph stack and also displays now/min/max/avg
values in the legend.